requirements.txt
asgiref==3.2.10
astroid==2.4.2
django-filter==2.3.0
django==3.0.8
djangorestframework==3.11.0
flake8==3.8.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
mypy==0.782
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
wrapt==1.12.1
